Fix recovering from malformed compiler list
A halt during compiler download can lead the compiler list to be partially written. On a subsequent run the parsing of the compiler list fails and leaves the hardhat install in a broken state.

We now attempt to parse the file and on `SyntaxError`, redownload and retry the parse once more. Other errors are rethrown.

Relates to #1466.