Modify download util to work via tmp file
An unexpected exit during download can leave a download file as corrupt. The download util now downloads to a temporary file (the filepath prefixed with `tmp-` so `compilers/tmp-list.json`) in the same directory as the intended download filepath. On completion of download the tmp file is moved into place at the expected filepath.

An unexpected exit will leave a tmp file, however the next attempted download will overwrite it and move it into place.